+++ Cargo.toml
@@ -20,3 +20,4 @@ tracing-subscriber = "0.3.19"
 im = { version = "15.1.0", features = ["serde"] }
 anyhow = "1.0.98"
 prost = "0.14.1"
+nix = { version = "0.30.1", features = ["user"] }
