+++ wire/key_agent/Cargo.toml
@@ -7,7 +7,7 @@ version.workspace = true
 tokio = { workspace = true }
 anyhow = { workspace = true }
 prost = { workspace = true }
-nix = { version = "0.30.1", features = ["user"] }
+nix = { workspace = true }
 
 [build-dependencies]
 prost-build = "0.14"
