+++ wire/lib/Cargo.toml
@@ -24,3 +24,4 @@ futures = "0.3.31"
 prost = { workspace = true }
 gethostname = "1.0.2"
 async-trait = "0.1.88"
+nix = { workspace = true }
